update students controller

+++ app/controllers/students_controller.rb
@@ -3,10 +3,15 @@
 class StudentsController < ApplicationController
   before_action :find_student, only: %i[update show destroy]
   before_action :authorized, except: %i[login]
+  before_action :user_roles_list, except: %i[login]
 
   def index
-    @students = Student.all.limit(10).to_json({ include: { roles: { include: 'permissions' } } })
-    json_response(@students, :created)
+    if (@user_roles_list & %w[Staff Admin]).any?
+      @students = Student.all.limit(10).to_json({ include: { roles: { include: 'permissions' } } })
+      json_response(@students, :created)
+    else
+      json_response('Not authorized', :unauthorized)
+    end
   end
 
   def login
@@ -22,41 +27,61 @@ class StudentsController < ApplicationController
   end
 
   def create
-    @student = Student.create(student_params)
-    @student.password = random_password
-    if @student.save
-      result = @student.to_json({ include: [:streams, { roles: { include: 'permissions' } }] })
-      json_response(result, :created)
+    if current_user.class.name == 'Principle' && @user_roles_list.include?('Admin')
+      @student = Student.create(student_params)
+      @student.password = random_password
+      if @student.save
+        result = @student.to_json({ include: [:streams, { roles: { include: 'permissions' } }] })
+        json_response(result, :created)
+      else
+        json_response('Error creating Student', :unprocessable_entity)
+      end
     else
-      json_response('Error creating Student', :unprocessable_entity)
+      json_response('Not authorized', :unauthorized)
     end
   end
 
   def show
-    result = @student.to_json({ include: [:streams, { roles: { include: 'permissions' } }] })
-    json_response(result, :created)
-  end
-
-  def update
-    if @student.update(student_params)
+    if (@user_roles_list & %w[Staff Admin Student]).any?
       result = @student.to_json({ include: [:streams, { roles: { include: 'permissions' } }] })
       json_response(result, :created)
     else
-      json_response('Cannot update Student', :unprocessable_entity)
+      json_response('Not authorized', :unauthorized)
+    end
+  end
+
+  def update
+    if (@user_roles_list & %w[Admin Student]).any?
+      if @student.update(student_params)
+        result = @student.to_json({ include: [:streams, { roles: { include: 'permissions' } }] })
+        json_response(result, :created)
+      else
+        json_response('Cannot update Student', :unprocessable_entity)
+      end
+    else
+      json_response('Not authorized', :unauthorized)
     end
   end
 
   def destroy
-    if @student.destroy
-      data = { 'message' => 'student destroyed' }
-      json_response(data, :no_content)
+    if current_user.class.name == 'Principle' && @user_roles_list.include?('Admin')
+      if @student.destroy
+        data = { 'message' => 'student destroyed' }
+        json_response(data, :no_content)
+      else
+        data = { 'message' => 'Something went wrong' }
+        json_response(data, :no_content)
+      end
     else
-      data = { 'message' => 'Something went wrong' }
-      json_response(data, :no_content)
+      json_response('Not authorized', :unauthorized)
     end
   end
 
   private
+
+  def user_roles_list
+    determine_type(current_user)
+  end
 
   def find_student
     @student = Student.find(params[:id])
@@ -68,3 +93,6 @@ class StudentsController < ApplicationController
     params.permit(:first_name, :last_name, :email, :password, role_ids: [], stream_ids: [])
   end
 end
+
+# johndoe@gmail.com
+# qSpbLK06Tf
